+++ spec/mailers/claim_mailer_spec.rb
@@ -1,8 +1,12 @@
 require "rails_helper"
 
-RSpec.shared_examples "a claim mailer" do
+RSpec.shared_examples "a claim mailer" do |policy|
   it "sets the correct to address" do
     expect(mail.to).to eq([claim.email_address])
+  end
+
+  it "sets the correct GOV.UK Notify reply to id" do
+    expect(mail["reply_to_id"].value).to eql(policy.notify_reply_to_id)
   end
 end
 
@@ -11,7 +15,7 @@ RSpec.describe ClaimMailer, type: :mailer do
     let(:claim) { create(:claim, :submittable, first_name: "Abraham", surname: "Lincoln") }
     let(:mail) { ClaimMailer.submitted(claim) }
 
-    it_behaves_like "a claim mailer"
+    it_behaves_like "a claim mailer", StudentLoans
 
     it "renders the subject" do
       expect(mail.subject).to eq("Your claim was received")
@@ -28,7 +32,7 @@ RSpec.describe ClaimMailer, type: :mailer do
     let(:claim) { create(:claim, :submitted, first_name: "John", middle_name: "Fitzgerald", surname: "Kennedy") }
     let(:mail) { ClaimMailer.approved(claim) }
 
-    it_behaves_like "a claim mailer"
+    it_behaves_like "a claim mailer", StudentLoans
 
     it "renders the subject" do
       expect(mail.subject).to match("approved")
@@ -45,7 +49,7 @@ RSpec.describe ClaimMailer, type: :mailer do
     let(:claim) { create(:claim, :submitted, first_name: "John", middle_name: "Fitzgerald", surname: "Kennedy") }
     let(:mail) { ClaimMailer.rejected(claim) }
 
-    it_behaves_like "a claim mailer"
+    it_behaves_like "a claim mailer", StudentLoans
 
     it "renders the subject" do
       expect(mail.subject).to match("rejected")
@@ -64,7 +68,7 @@ RSpec.describe ClaimMailer, type: :mailer do
     let(:payment_date_timestamp) { Time.new(2019, 1, 1).to_i }
     let(:mail) { ClaimMailer.payment_confirmation(payment.claim, payment_date_timestamp) }
 
-    it_behaves_like "a claim mailer"
+    it_behaves_like "a claim mailer", StudentLoans
 
     it "renders the subject" do
       expect(mail.subject).to match("paying")